requirements.txt
pydantic==2.11.5
fastapi==0.115.12
uvicorn==0.34.3
gunicorn==23.0.0
openai==1.86.0
pymongo==4.13.1
backoff==2.2.1
aiohttp==3.12.13
redis==6.2.0
motor==3.7.1
websockets==15.0.1
python-dotenv==1.1.0